Extract client App to separate file (easier 2 test)
